requirements.txt
Flask
Werkzeug
waitress
Pillow
opencv-python-headless
numpy
